Fix all suggestions from rider in test project
