+++ backend/tests/Services/DrainageStrategyTimeSeriesServiceTests.cs
@@ -9,87 +9,86 @@ using NSubstitute;
 
 using Xunit;
 
-namespace tests.Services
+namespace tests.Services;
+
+public class DrainageStrategyTimeSeriesServiceTests
 {
-    public class DrainageStrategyTimeSeriesServiceTests
+    private readonly IDrainageStrategyTimeSeriesService _drainageStrategyTimeSeriesService;
+    private readonly ILogger<DrainageStrategyService> _logger = Substitute.For<ILogger<DrainageStrategyService>>();
+    private readonly IDrainageStrategyTimeSeriesRepository _repository = Substitute.For<IDrainageStrategyTimeSeriesRepository>();
+    private readonly IDrainageStrategyRepository _drainageStrategyRepository = Substitute.For<IDrainageStrategyRepository>();
+    private readonly ICaseRepository _caseRepository = Substitute.For<ICaseRepository>();
+    private readonly IConversionMapperService _conversionMapperService = Substitute.For<IConversionMapperService>();
+    private readonly IProjectRepository _projectRepository = Substitute.For<IProjectRepository>();
+    private readonly IProjectAccessService _projectAccessService = Substitute.For<IProjectAccessService>();
+
+    public DrainageStrategyTimeSeriesServiceTests()
     {
-        private readonly IDrainageStrategyTimeSeriesService _drainageStrategyTimeSeriesService;
-        private readonly ILogger<DrainageStrategyService> _logger = Substitute.For<ILogger<DrainageStrategyService>>();
-        private readonly IDrainageStrategyTimeSeriesRepository _repository = Substitute.For<IDrainageStrategyTimeSeriesRepository>();
-        private readonly IDrainageStrategyRepository _drainageStrategyRepository = Substitute.For<IDrainageStrategyRepository>();
-        private readonly ICaseRepository _caseRepository = Substitute.For<ICaseRepository>();
-        private readonly IConversionMapperService _conversionMapperService = Substitute.For<IConversionMapperService>();
-        private readonly IProjectRepository _projectRepository = Substitute.For<IProjectRepository>();
-        private readonly IProjectAccessService _projectAccessService = Substitute.For<IProjectAccessService>();
+        _drainageStrategyTimeSeriesService = new DrainageStrategyTimeSeriesService(
+            _logger,
+            _caseRepository,
+            _repository,
+            _drainageStrategyRepository,
+            _conversionMapperService,
+            _projectRepository,
+            _projectAccessService
+        );
+    }
 
-        public DrainageStrategyTimeSeriesServiceTests()
-        {
-            _drainageStrategyTimeSeriesService = new DrainageStrategyTimeSeriesService(
-                _logger,
-                _caseRepository,
-                _repository,
-                _drainageStrategyRepository,
-                _conversionMapperService,
-                _projectRepository,
-                _projectAccessService
-            );
-        }
+    [Fact]
+    public async Task CreateDeferredGasProduction_ShouldReturnDto_WhenSuccessful()
+    {
+        // Arrange
+        var projectId = Guid.NewGuid();
+        var caseId = Guid.NewGuid();
+        var drainageStrategyId = Guid.NewGuid();
+        var createProfileDto = new CreateDeferredGasProductionDto();
 
-        [Fact]
-        public async Task CreateDeferredGasProduction_ShouldReturnDto_WhenSuccessful()
-        {
-            // Arrange
-            var projectId = Guid.NewGuid();
-            var caseId = Guid.NewGuid();
-            var drainageStrategyId = Guid.NewGuid();
-            var createProfileDto = new CreateDeferredGasProductionDto();
+        var drainageStrategy = new DrainageStrategy();
+        _drainageStrategyRepository.GetDrainageStrategy(drainageStrategyId).Returns(drainageStrategy);
 
-            var drainageStrategy = new DrainageStrategy();
-            _drainageStrategyRepository.GetDrainageStrategy(drainageStrategyId).Returns(drainageStrategy);
+        _drainageStrategyRepository.DrainageStrategyHasProfile(drainageStrategyId, DrainageStrategyProfileNames.DeferredGasProduction).Returns(false);
 
-            _drainageStrategyRepository.DrainageStrategyHasProfile(drainageStrategyId, DrainageStrategyProfileNames.DeferredGasProduction).Returns(false);
+        var project = new Project();
+        _projectRepository.GetProject(projectId).Returns(project);
 
-            var project = new Project();
-            _projectRepository.GetProject(projectId).Returns(project);
+        var createdProfile = new DeferredGasProduction();
+        _repository.CreateDeferredGasProduction(Arg.Any<DeferredGasProduction>()).Returns(createdProfile);
 
-            var createdProfile = new DeferredGasProduction();
-            _repository.CreateDeferredGasProduction(Arg.Any<DeferredGasProduction>()).Returns(createdProfile);
+        var expectedDto = new DeferredGasProductionDto();
+        _conversionMapperService.MapToDto<DeferredGasProduction, DeferredGasProductionDto>(createdProfile, createdProfile.Id, project.PhysicalUnit).Returns(expectedDto);
 
-            var expectedDto = new DeferredGasProductionDto();
-            _conversionMapperService.MapToDto<DeferredGasProduction, DeferredGasProductionDto>(createdProfile, createdProfile.Id, project.PhysicalUnit).Returns(expectedDto);
+        // Act
+        var result = await _drainageStrategyTimeSeriesService.CreateDeferredGasProduction(projectId, caseId, drainageStrategyId, createProfileDto);
 
-            // Act
-            var result = await _drainageStrategyTimeSeriesService.CreateDeferredGasProduction(projectId, caseId, drainageStrategyId, createProfileDto);
+        // Assert
+        Assert.Equal(expectedDto, result);
+    }
 
-            // Assert
-            Assert.Equal(expectedDto, result);
-        }
+    [Fact]
+    public async Task CreateDeferredGasProduction_ShouldThrow_WhenResourceExists()
+    {
+        // Arrange
+        var projectId = Guid.NewGuid();
+        var caseId = Guid.NewGuid();
+        var drainageStrategyId = Guid.NewGuid();
+        var createProfileDto = new CreateDeferredGasProductionDto();
 
-        [Fact]
-        public async Task CreateDeferredGasProduction_ShouldThrow_WhenResourceExists()
-        {
-            // Arrange
-            var projectId = Guid.NewGuid();
-            var caseId = Guid.NewGuid();
-            var drainageStrategyId = Guid.NewGuid();
-            var createProfileDto = new CreateDeferredGasProductionDto();
+        var drainageStrategy = new DrainageStrategy();
+        _drainageStrategyRepository.GetDrainageStrategy(drainageStrategyId).Returns(drainageStrategy);
 
-            var drainageStrategy = new DrainageStrategy();
-            _drainageStrategyRepository.GetDrainageStrategy(drainageStrategyId).Returns(drainageStrategy);
+        _drainageStrategyRepository.DrainageStrategyHasProfile(drainageStrategyId, DrainageStrategyProfileNames.DeferredGasProduction).Returns(true);
 
-            _drainageStrategyRepository.DrainageStrategyHasProfile(drainageStrategyId, DrainageStrategyProfileNames.DeferredGasProduction).Returns(true);
+        var project = new Project();
+        _projectRepository.GetProject(projectId).Returns(project);
 
-            var project = new Project();
-            _projectRepository.GetProject(projectId).Returns(project);
+        var createdProfile = new DeferredGasProduction();
+        _repository.CreateDeferredGasProduction(Arg.Any<DeferredGasProduction>()).Returns(createdProfile);
 
-            var createdProfile = new DeferredGasProduction();
-            _repository.CreateDeferredGasProduction(Arg.Any<DeferredGasProduction>()).Returns(createdProfile);
+        var expectedDto = new DeferredGasProductionDto();
+        _conversionMapperService.MapToDto<DeferredGasProduction, DeferredGasProductionDto>(createdProfile, createdProfile.Id, project.PhysicalUnit).Returns(expectedDto);
 
-            var expectedDto = new DeferredGasProductionDto();
-            _conversionMapperService.MapToDto<DeferredGasProduction, DeferredGasProductionDto>(createdProfile, createdProfile.Id, project.PhysicalUnit).Returns(expectedDto);
-
-            // Act & Assert
-            await Assert.ThrowsAsync<ResourceAlreadyExistsException>(() => _drainageStrategyTimeSeriesService.CreateDeferredGasProduction(projectId, caseId, drainageStrategyId, createProfileDto));
-        }
+        // Act & Assert
+        await Assert.ThrowsAsync<ResourceAlreadyExistsException>(() => _drainageStrategyTimeSeriesService.CreateDeferredGasProduction(projectId, caseId, drainageStrategyId, createProfileDto));
     }
 }
