+++ backend/tests/Services/SurfServiceTimeSeriesTests.cs
@@ -7,125 +7,124 @@ using NSubstitute;
 
 using Xunit;
 
-namespace tests.Services
+namespace tests.Services;
+
+public class SurfServiceTimeSeriesTests
 {
-    public class SurfServiceTimeSeriesTests
+    private readonly SurfTimeSeriesService _surfService;
+    private readonly ILogger<SurfService> _logger = Substitute.For<ILogger<SurfService>>();
+    private readonly ISurfTimeSeriesRepository _repository = Substitute.For<ISurfTimeSeriesRepository>();
+    private readonly ISurfRepository _surfRepository = Substitute.For<ISurfRepository>();
+    private readonly ICaseRepository _caseRepository = Substitute.For<ICaseRepository>();
+    private readonly IMapperService _mapperService = Substitute.For<IMapperService>();
+    private readonly IProjectAccessService _projectAccessService = Substitute.For<IProjectAccessService>();
+
+    public SurfServiceTimeSeriesTests()
     {
-        private readonly SurfTimeSeriesService _surfService;
-        private readonly ILogger<SurfService> _logger = Substitute.For<ILogger<SurfService>>();
-        private readonly ISurfTimeSeriesRepository _repository = Substitute.For<ISurfTimeSeriesRepository>();
-        private readonly ISurfRepository _surfRepository = Substitute.For<ISurfRepository>();
-        private readonly ICaseRepository _caseRepository = Substitute.For<ICaseRepository>();
-        private readonly IMapperService _mapperService = Substitute.For<IMapperService>();
-        private readonly IProjectAccessService _projectAccessService = Substitute.For<IProjectAccessService>();
+        _surfService = new SurfTimeSeriesService(
+            _logger,
+            _repository,
+            _surfRepository,
+            _caseRepository,
+            _mapperService,
+            _projectAccessService
+        );
+    }
 
-        public SurfServiceTimeSeriesTests()
+    [Fact]
+    public async Task UpdateSurfCostProfileOverride_ShouldUpdateSurfCostProfileOverride_WhenGivenValidInput()
+    {
+        // Arrange
+        var projectId = Guid.NewGuid();
+        var caseId = Guid.NewGuid();
+        var surfId = Guid.NewGuid();
+        var costProfileId = Guid.NewGuid();
+        var updatedSurfCostProfileOverrideDto = new UpdateSurfCostProfileOverrideDto();
+
+        var existingSurfCostProfileOverride = new SurfCostProfileOverride
         {
-            _surfService = new SurfTimeSeriesService(
-                _logger,
-                _repository,
-                _surfRepository,
-                _caseRepository,
-                _mapperService,
-                _projectAccessService
-            );
-        }
+            Id = costProfileId,
+            Surf = new Surf { Id = surfId }
+        };
+        _repository.GetSurfCostProfileOverride(costProfileId).Returns(existingSurfCostProfileOverride);
 
-        [Fact]
-        public async Task UpdateSurfCostProfileOverride_ShouldUpdateSurfCostProfileOverride_WhenGivenValidInput()
+        var updatedSurfCostProfileOverride = new SurfCostProfileOverride
         {
-            // Arrange
-            var projectId = Guid.NewGuid();
-            var caseId = Guid.NewGuid();
-            var surfId = Guid.NewGuid();
-            var costProfileId = Guid.NewGuid();
-            var updatedSurfCostProfileOverrideDto = new UpdateSurfCostProfileOverrideDto();
+            Id = costProfileId,
+            Surf = new Surf { Id = surfId }
+        };
+        _repository.UpdateSurfCostProfileOverride(existingSurfCostProfileOverride).Returns(updatedSurfCostProfileOverride);
 
-            var existingSurfCostProfileOverride = new SurfCostProfileOverride
-            {
-                Id = costProfileId,
-                Surf = new Surf { Id = surfId }
-            };
-            _repository.GetSurfCostProfileOverride(costProfileId).Returns(existingSurfCostProfileOverride);
+        var updatedSurfCostProfileOverrideDtoResult = new SurfCostProfileOverrideDto();
+        _mapperService.MapToDto<SurfCostProfileOverride, SurfCostProfileOverrideDto>(existingSurfCostProfileOverride, costProfileId).Returns(updatedSurfCostProfileOverrideDtoResult);
 
-            var updatedSurfCostProfileOverride = new SurfCostProfileOverride
-            {
-                Id = costProfileId,
-                Surf = new Surf { Id = surfId }
-            };
-            _repository.UpdateSurfCostProfileOverride(existingSurfCostProfileOverride).Returns(updatedSurfCostProfileOverride);
+        // Act
+        var result = await _surfService.UpdateSurfCostProfileOverride(projectId, caseId, surfId, costProfileId, updatedSurfCostProfileOverrideDto);
 
-            var updatedSurfCostProfileOverrideDtoResult = new SurfCostProfileOverrideDto();
-            _mapperService.MapToDto<SurfCostProfileOverride, SurfCostProfileOverrideDto>(existingSurfCostProfileOverride, costProfileId).Returns(updatedSurfCostProfileOverrideDtoResult);
+        // Assert
+        Assert.Equal(updatedSurfCostProfileOverrideDtoResult, result);
+        await _repository.Received(1).SaveChangesAndRecalculateAsync(caseId);
+    }
 
-            // Act
-            var result = await _surfService.UpdateSurfCostProfileOverride(projectId, caseId, surfId, costProfileId, updatedSurfCostProfileOverrideDto);
+    [Fact]
+    public async Task AddOrUpdateSurfCostProfile_ShouldUpdateSurfCostProfile_WhenGivenValidInputForExistingProfile()
+    {
+        // Arrange
+        var projectId = Guid.NewGuid();
+        var caseId = Guid.NewGuid();
+        var surfId = Guid.NewGuid();
+        var profileId = Guid.NewGuid();
+        var updatedSurfCostProfileDto = new UpdateSurfCostProfileDto();
 
-            // Assert
-            Assert.Equal(updatedSurfCostProfileOverrideDtoResult, result);
-            await _repository.Received(1).SaveChangesAndRecalculateAsync(caseId);
-        }
-
-        [Fact]
-        public async Task AddOrUpdateSurfCostProfile_ShouldUpdateSurfCostProfile_WhenGivenValidInputForExistingProfile()
+        var existingCostProfile = new SurfCostProfile
         {
-            // Arrange
-            var projectId = Guid.NewGuid();
-            var caseId = Guid.NewGuid();
-            var surfId = Guid.NewGuid();
-            var profileId = Guid.NewGuid();
-            var updatedSurfCostProfileDto = new UpdateSurfCostProfileDto();
+            Id = profileId,
+            Surf = new Surf { Id = surfId }
+        };
+        var existingSurf = new Surf { Id = surfId, CostProfile = existingCostProfile };
+        _surfRepository.GetSurfWithCostProfile(surfId).Returns(existingSurf);
 
-            var existingCostProfile = new SurfCostProfile
-            {
-                Id = profileId,
-                Surf = new Surf { Id = surfId }
-            };
-            var existingSurf = new Surf { Id = surfId, CostProfile = existingCostProfile };
-            _surfRepository.GetSurfWithCostProfile(surfId).Returns(existingSurf);
+        _repository.GetSurfCostProfile(profileId).Returns(existingCostProfile);
+        _repository.UpdateSurfCostProfile(existingCostProfile).Returns(existingCostProfile);
 
-            _repository.GetSurfCostProfile(profileId).Returns(existingCostProfile);
-            _repository.UpdateSurfCostProfile(existingCostProfile).Returns(existingCostProfile);
+        var updatedSurfCostProfileDtoResult = new SurfCostProfileDto { Id = profileId };
+        _mapperService.MapToDto<SurfCostProfile, SurfCostProfileDto>(existingCostProfile, existingCostProfile.Id).Returns(updatedSurfCostProfileDtoResult);
 
-            var updatedSurfCostProfileDtoResult = new SurfCostProfileDto { Id = profileId };
-            _mapperService.MapToDto<SurfCostProfile, SurfCostProfileDto>(existingCostProfile, existingCostProfile.Id).Returns(updatedSurfCostProfileDtoResult);
+        // Act
+        var result = await _surfService.AddOrUpdateSurfCostProfile(projectId, caseId, surfId, updatedSurfCostProfileDto);
 
-            // Act
-            var result = await _surfService.AddOrUpdateSurfCostProfile(projectId, caseId, surfId, updatedSurfCostProfileDto);
+        // Assert
+        Assert.Equal(updatedSurfCostProfileDtoResult, result);
+        await _repository.Received(1).SaveChangesAndRecalculateAsync(caseId);
+    }
 
-            // Assert
-            Assert.Equal(updatedSurfCostProfileDtoResult, result);
-            await _repository.Received(1).SaveChangesAndRecalculateAsync(caseId);
-        }
+    [Fact]
+    public async Task AddOrUpdateSurfCostProfile_ShouldAddSurfCostProfile_WhenGivenValidInputForNewProfile()
+    {
+        // Arrange
+        var projectId = Guid.NewGuid();
+        var caseId = Guid.NewGuid();
+        var surfId = Guid.NewGuid();
+        var profileId = Guid.NewGuid();
+        var updatedSurfCostProfileDto = new UpdateSurfCostProfileDto();
 
-        [Fact]
-        public async Task AddOrUpdateSurfCostProfile_ShouldAddSurfCostProfile_WhenGivenValidInputForNewProfile()
-        {
-            // Arrange
-            var projectId = Guid.NewGuid();
-            var caseId = Guid.NewGuid();
-            var surfId = Guid.NewGuid();
-            var profileId = Guid.NewGuid();
-            var updatedSurfCostProfileDto = new UpdateSurfCostProfileDto();
+        var existingSurf = new Surf { Id = surfId };
+        _surfRepository.GetSurfWithCostProfile(surfId).Returns(existingSurf);
 
-            var existingSurf = new Surf { Id = surfId };
-            _surfRepository.GetSurfWithCostProfile(surfId).Returns(existingSurf);
+        var newCostProfile = new SurfCostProfile { Surf = existingSurf };
+        _mapperService.MapToEntity(Arg.Any<UpdateSurfCostProfileDto>(), Arg.Any<SurfCostProfile>(), Arg.Any<Guid>())
+            .Returns(newCostProfile);
 
-            var newCostProfile = new SurfCostProfile { Surf = existingSurf };
-            _mapperService.MapToEntity(Arg.Any<UpdateSurfCostProfileDto>(), Arg.Any<SurfCostProfile>(), Arg.Any<Guid>())
-                          .Returns(newCostProfile);
+        _repository.CreateSurfCostProfile(newCostProfile).Returns(newCostProfile);
 
-            _repository.CreateSurfCostProfile(newCostProfile).Returns(newCostProfile);
+        var updatedSurfCostProfileDtoResult = new SurfCostProfileDto { Id = profileId };
+        _mapperService.MapToDto<SurfCostProfile, SurfCostProfileDto>(newCostProfile, newCostProfile.Id).Returns(updatedSurfCostProfileDtoResult);
 
-            var updatedSurfCostProfileDtoResult = new SurfCostProfileDto { Id = profileId };
-            _mapperService.MapToDto<SurfCostProfile, SurfCostProfileDto>(newCostProfile, newCostProfile.Id).Returns(updatedSurfCostProfileDtoResult);
+        // Act
+        var result = await _surfService.AddOrUpdateSurfCostProfile(projectId, caseId, surfId, updatedSurfCostProfileDto);
 
-            // Act
-            var result = await _surfService.AddOrUpdateSurfCostProfile(projectId, caseId, surfId, updatedSurfCostProfileDto);
-
-            // Assert
-            Assert.Equal(updatedSurfCostProfileDtoResult, result);
-            await _repository.Received(1).SaveChangesAndRecalculateAsync(caseId);
-        }
+        // Assert
+        Assert.Equal(updatedSurfCostProfileDtoResult, result);
+        await _repository.Received(1).SaveChangesAndRecalculateAsync(caseId);
     }
 }
